audio_video_to_text/requirements.txt
vertexai
python-dotenv
ffmpeg-python